Refactor get_broken_links to remove file access
get_broken_links() was hitting the file system in order to
retrieve the entire list of vanilla sound events. This change
makes the function slightly more testable. Now I just need
to remove the reference to symlinks.

Also moved the definition of assets_folder so that I can
use it in the call to get ogg files.

Contributes to issue #10 (tangentially)