Add tests for get_real_path
Along the way, I felt the need to rename the existing "test" folder to
"test-environment" to differentiate it a bit from the new "tests"
folder.

While I was in the mood to rename things, I renamed the script too,
opting for a much shorter "spcheck.py" so that when I'm going between this
and the indexer, it only takes three characters to arrive at a
difference rather than having to type "sound-pack-i" to get a different
auto-complete. Besides that, I read somewhere that a best practice for
module names is really short single-word names.

Contributes to issue #10