Move color codes back to local variables
I wasn't happy with how much text was required to use an enum in Python,
so I went back to a more local way of defining those.  This doesn't
contribute to any issue in particular, I just thought of it while at
work last night and wanted to do it.

... also cleaned up some imports that apparently weren't in use

Chalk it up to "tech debt," maybe?