Add a test to return broken symlinks
If a sounds.json link to a file points to a valid path, but the path is
a symlink that doesn't point to anything, that should show up on the
list of broken links returned from the function. This works, but I
didn't have a test for it.

Contributes to issue #10