Refactor get_irrelevant_files function
Now uses the new "all_files" list to determine which files (if any) are
not supposed to be in the pack folder structure.

Also moved the call to the function to the top, so that if files are
declared irrelevant, they can be removed from "all_files" and not
considered in the next step.

Added a new test to verify that the resulting list is sorted
alphabetically.

Contributes to issue #10