Refactor get_orphaned_files function
Moved it to the second step, because it seemed like the most logical
next step to eliminate bad files from my new list of files.

Added two new tests to validate symlink behavior - files that are
referenced by JSON through a symlink shouldn't be considered "orphans"
because they have something that points to them - files that are broken
symlinks referenced by JSON shouldn't be considered "orphans" because
they will not exist once the package is compressed into a zip file.
Besides that, the "broken links" step should have removed those for us
already anyway (and I have tests for that.)

Orphans are now removed from the list of ogg_files.

Changed a few comments to try to provide more clarity.

Contributes to issue #10