Refactor get_broken_links and add Path subclass
Added a new subclass of Path so that I can check whether a file is a
symlink as soon as its path is known, and resolve its target folder as
well. This allows me to check these things down the line without
needing to hit the file system later (or use a fake file system to test it.)

A lot of things changed during this commit. It was brutal; taking much
longer than I thought to complete. There's more to do, but I wanted to
get this one checked in because it was already getting huge. It may turn
out to be a huge mistake that I'll need to revert... who knows?

I added a few tests that came up while making the existing tests work, I
changed all references to Path to a custom version with extra
properties (probably should have renamed it, but PyCharm decided that it
couldn't rename anything that wasn't in the root for some reason.)

Changed the order of a file check in get_real_path because I was getting an
error message that didn't make sense in the context of what I was
testing.

Contributes to issue #10 (I think...)