Refactor get_invalid_file_names
Moved it to the last position (among the functions that gather bad file
paths) so that it matches its order in main() and changed the argument
to ogg_files rather than the events dictionary. With a list of files
rather than a dictionary, the logic is a lot more straightforward.

Also changed the statements that strip bad files out of the ogg_files
list so that the list is getting smaller each time. By the time it makes
it to get_invalid_file_names, the list should have been reduced to
those items that work or have invalid file names.

Had to change the tests for that function as well. One of them worries
me, since it was passing even with the dictionary of sound events being
passed in rather than a list of ogg files.

Contributes to issue #10