Refactor get_broken_links and add some tests
Removed the need to gather vanilla sounds every time through the loop
for JSON events. This reduces nesting and makes the logic easier to
understand.

Added two tests - one to check multiple vanilla sounds in the same test,
and another to verify that the results are being returned after being
sorted alphabetically.

Contributes to issue #10