Add SoundEventHandler object and tests
Massive changes that I didn't expect. Nearly every file in the project
has been modified. I've added an object to handle all the json records
in the pack. It stores the "root" folder (which is basically "assets"
and anything that leads up to it.) It does path formatting for
comparisons, and makes sure that all the lists that it outputs are
sorted alphabetically.

I may want to drop symlink support in this app, because that added a ton
of complexity to the logic that I probably don't need anymore if I just
say "you can't use symlinks in a sound pack anymore."  In fact, I'm not
even sure why I started using them in the first place.

I think I got it working the way it used to, but there may be a bunch of
errors I haven't come across yet.

I feel like a change this massive deserves an explanation of what
changed, but so much changed that I have no idea how I would cover it
all.

Contributes to issue #10